Move java files
