+++ database/app/src/main/AndroidManifest.xml
@@ -9,18 +9,18 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".MainActivity"
+            android:name=".java.MainActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme" />
-        <activity android:name=".NewPostActivity" />
-        <activity android:name=".SignInActivity">
+        <activity android:name=".java.NewPostActivity" />
+        <activity android:name=".java.SignInActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".PostDetailActivity"></activity>
+        <activity android:name=".java.PostDetailActivity"></activity>
     </application>
 
 </manifest>
