+++ database/app/src/main/java/com/google/firebase/quickstart/database/java/PostDetailActivity.java
@@ -1,4 +1,4 @@
-package com.google.firebase.quickstart.database;
+package com.google.firebase.quickstart.database.java;
 
 import android.content.Context;
 import android.os.Bundle;
@@ -19,9 +19,10 @@ import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
-import com.google.firebase.quickstart.database.models.User;
-import com.google.firebase.quickstart.database.models.Comment;
-import com.google.firebase.quickstart.database.models.Post;
+import com.google.firebase.quickstart.database.R;
+import com.google.firebase.quickstart.database.java.models.User;
+import com.google.firebase.quickstart.database.java.models.Comment;
+import com.google.firebase.quickstart.database.java.models.Post;
 
 import java.util.ArrayList;
 import java.util.List;
