+++ database/app/src/main/java/com/google/firebase/quickstart/database/java/SignInActivity.java
@@ -1,4 +1,4 @@
-package com.google.firebase.quickstart.database;
+package com.google.firebase.quickstart.database.java;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -17,7 +17,8 @@ import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.quickstart.database.models.User;
+import com.google.firebase.quickstart.database.R;
+import com.google.firebase.quickstart.database.java.models.User;
 
 public class SignInActivity extends BaseActivity implements View.OnClickListener {
