+++ database/app/src/main/res/layout/activity_post_detail.xml
@@ -6,8 +6,7 @@
     android:paddingBottom="@dimen/activity_vertical_margin"
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.google.firebase.quickstart.database.PostDetailActivity">
+    android:paddingTop="@dimen/activity_vertical_margin">
 
     <include
         android:id="@+id/post_author_layout"
