+++ database/app/src/main/res/layout/activity_sign_in.xml
@@ -6,8 +6,7 @@
     android:paddingBottom="@dimen/activity_vertical_margin"
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.google.firebase.quickstart.database.SignInActivity">
+    android:paddingTop="@dimen/activity_vertical_margin">
 
 
     <ImageView
